Tests/LibWeb/Text/expected/HTML/img-metrics-in-load-event-handlers.txt

image is 400x400
